short to int solution num

+++ src/main/core/MacroState.java
@@ -8,24 +8,24 @@ import java.util.List;
  */
 public class MacroState {
 
-    private List<Short> solutions;
+    private List<Integer> solutions;
 
-    private int solutionsEvaluation;
+    private long solutionsEvaluation;
 
-    public MacroState(int solutionsEvaluation) {
+    public MacroState(long solutionsEvaluation) {
         this.solutionsEvaluation = solutionsEvaluation;
-        solutions = new LinkedList<Short>();
+        solutions = new LinkedList<Integer>();
     }
 
-    public void addSolutionToMacroState(Short solutionNumber){
+    public void addSolutionToMacroState(Integer solutionNumber){
         solutions.add(solutionNumber);
     }
 
-    public List<Short> getSolutions() {
+    public List<Integer> getSolutions() {
         return solutions;
     }
 
-    public int getSolutionsEvaluation() {
+    public long getSolutionsEvaluation() {
         return solutionsEvaluation;
     }
 }
